Address MSVC++ compiler issues; mostly missing header file includes.
git-svn-id: https://svnsrv.desy.de/basic/PandoraPFANew/PandoraPFANew/trunk@584 e861a63a-b6b5-47dc-a5de-b4304083765c

+++ src/Helpers/CaloHitHelper.cc
@@ -114,7 +114,7 @@ float CaloHitHelper::GetDensityWeightContribution(const CaloHit *const pCaloHit,
 {
     static const float caloHitMaxSeparation(PandoraSettings::GetInstance()->GetCaloHitMaxSeparation());
     static const float caloHitMaxSeparationSquared(caloHitMaxSeparation * caloHitMaxSeparation);
-    static const float densityWeightPower(PandoraSettings::GetInstance()->GetDensityWeightPower());
+    static const unsigned int densityWeightPower(PandoraSettings::GetInstance()->GetDensityWeightPower());
 
     float densityWeightContribution = 0.;
     const CartesianVector &positionVector(pCaloHit->GetPositionVector());
